Simplification of SplitPointHandler
We don't need the findLeading, skipLeading and trimLeading
functions. These were used for processing skippable rows at the
beginning of a page, but if we make sure that SplitPointHandler.find()
always skips skippable rows at the end of a page, we're also good.

In fact this fixes a funny behavior that existed in Dotify, namely
that bottom margins were also skipped if they were present at the top
of the page (which could happen in case of an empty block with a
margin-bottom).

Some code was added to keep the old behavior of marker-reference scope
"page-content" with any-resumed or sequence-resumed content present on
a page. Although this behavior is also not documented and perhaps
illogical, I decided to keep it for now because it is currently used
by DAISY Pipeline.