Don't ignore anchors in empty blocks
For this we needed to make empty blocks non-skippable again, in other
words this reverses commit
https://github.com/mtmse/dotify.formatter.impl/commit/c274122a0a. The
original reason for doing commit c274122a0a is not important anymore.

Note that this fixes only the page-area case. Anchors in empty blocks
are still ignored when the collection is used for a
list-of-references. A test shows this.