Fix another bug that resulted the dreaded "Failed to complete volume division" error
The "row count" data stored in CrossReferenceHandler is used by the
page break algorithm in the next iteration to determine/estimate how
many widows lines a block will have. The actual row count can differ
however, for instance due to changes in hyphenation (hyphenation may
be suppressed on the last line of the page). This could lead to an
endless alternation between two renderings. This is why I've changed
it to store the *minimum* value of all iterations. Note that it could
result in page breaks that happen one row too early because the row
count was underestimated.